+++ Robust.Client/UserInterface/Controls/MultiselectOptionButton.cs
@@ -5,6 +5,7 @@ using System.Numerics;
 using Robust.Client.Graphics;
 using Robust.Shared.Graphics;
 using Robust.Shared.Maths;
+using Robust.Shared.ViewVariables;
 using static Robust.Client.UserInterface.Controls.BoxContainer;
 
 namespace Robust.Client.UserInterface.Controls
@@ -57,6 +58,20 @@ namespace Robust.Client.UserInterface.Controls
         {
             get => _label.Text;
             set => _label.Text = value;
+        }
+
+        // Compatibility shim for old XAML behaviour that assumed setting
+        // classes would concatenate with the StyleClassButton instead of overwriting
+        // all of them
+        [ViewVariables]
+        new public StyleClassCollection StyleClasses
+        {
+            get => base.StyleClasses;
+            set
+            {
+                base.StyleClasses = value;
+                base.StyleClasses.Add(StyleClassButton);
+            }
         }
 
         public MultiselectOptionButton()